requirements.txt
mecab-python3==0.996.1
selenium==3.141.0
urllib3==1.24.1
